requirements.txt
appdirs==1.4.3
appnope==0.1.0
argcomplete==1.10.0
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.8.0
black==19.10b0
bleach==3.1.5
blis==0.4.1
boto==2.49.0
boto3==1.14.25
botocore==1.17.25
catalogue==1.0.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
docx2txt==0.8
EbookLib==0.17.1
entrypoints==0.3
extract-msg==0.23.1
Flask==1.1.1
future==0.18.2
gensim==3.8.3
gunicorn==20.0.4
idna==2.8
IMAPClient==2.1.0
ipykernel==5.3.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.10.3
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
lxml==4.5.2
MarkupSafe==1.1.1
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.5
notebook==6.0.3
numpy==1.19.0
olefile==0.46
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pathspec==0.6.0
pdfminer.six==20181108
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.19
pycryptodome==3.9.8
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-pptx==0.6.18
python-telegram-bot==12.2.0
pytz==2020.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
redis==3.5.3
regex==2019.11.1
requests==2.22.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.5.1
Send2Trash==1.5.0
six==1.12.0
smart-open==2.1.0
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==2.3.2
SpeechRecognition==3.8.1
srsly==1.0.2
terminado==0.8.3
testpath==0.4.4
textract==1.6.3
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.0
tornado==6.0.3
tqdm==4.48.0
traitlets==4.3.3
typed-ast==1.4.0
tzlocal==1.5.1
urllib3==1.25.7
wasabi==0.7.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
xlrd==1.2.0
XlsxWriter==1.2.9
